examples/requirements.txt
ipykernel==6.29.0
ipython==8.20.0
matplotlib==3.8.2
numpy==1.26.3
pygame-ce==2.4.0
